Added dependency injection for the Upload process (which affects most of the framework).
An example of what is shown is replacing web3.js with ethers.js which could be achieved through a config file setting. The abstraction of which still needs to be fleshed out, but the architecture of the dependency injection is there.

This is not fully functional. The plugins are not loading correctly due to Config and Plugins clasess being too tightly coupled for IOC. This should be refactored.

Test examples coming next.