Add support for dependency injection in Events and Logger.
This is for an DI MVP for just the Events and Logger objects in Embark. There are quite a few places where DI could not used in this limited scope because other options besides Events/Logger need to be passed in as well and the container would not be able to resolve all of them.

Apart from the above, if we need to get events or logger from anywhere in the application, we can use `container.get(Events)` or `container.get(Logger)`.