Finally fixed the solc loading issue.
Moved the instantiation of the live-plugin-manager in to the child solc process. This allowed us to use the live-plugin-manager to require the installed solc file.

If the module is the same version as used by embark, the module is still loading using the normal require (from node_modules), as before.