Added the queue back in, added cache reset
The nonce cache is incremented as contracts are asked to deploy. The nonce request is queued to avoid issues with concurrency.

The nonce cache is decremented when contract deploy receipt is received.

Starting at cache value of 1 seemed to alleviate some issues with “nonce too low” or “underpriced” errors.

Debugging console logs have been left in to help with illustrating the nonce cache increment and decrement. They should be removed prior to PR.