feature(@embark/core): IPC adapts when run in Docker Linux on Win
When embark is running on Linux and macOS, unix socket files are used for
geth's and embark's IPC files. If the context is a Linux container running on a
Windows Docker host, and if the DApp is mounted from the host's file system,
there is a problem: unix socket files are incompatible with the Windows file
system.

Detect the problematic scenario and use a path outside of the mounted directory
for embark's and geth's IPC files.

To avoid a circular dependency problem, remove the `utils/utils.js` dependency
from `core/env.js`, don't have `utils/host.ts` depdend on `core/fs.js`, and
implement the `ipcPath()` helper as a static method on the class exported from
`core/ipc.js`.