fix(@embark/core): Show transaction log for errors
When a transaction errors in the cockpit, show the error in the transaction log.

Contract logs were not occuring in the test_app for reasons unknown, except that it seemed to be pinned to the fact that accounts were being defined using a mnemonic. Turns out this caused transactions to be signed, and therefore sent using `eth_sendRawTransaction`.

This does not catch errors in web3 that do not make it server-side (ie when executing a tx from the console or from the browser).

Add ability to decode raw transactions in the proxy so that those transactions are also logged.

Additional updates:

Refactor proxy as a class. Add unit tests for Proxy class.

Add constants for transaction status and update references in the codebase where possible.

Fix proxy class references.

Prevent automatic debugging if no tx hash.

Add unit test for console log errors.