feat(@embark/embark-blockchain-connector): Add support for `eth_signTypedData_v3`
This PR adds support for EIP-712, by allowing the signing of typed data transactions using the `eth_signTypeData_v3` or `eth_signTypedData` request.

NOTE: In Embark’s current state, this is only applicable in cases where Embark intercepts the send (ie when the web3 provider is set to Embark’s provider). For example, this works when executing console commands that require signing typed data txs. However, if the situation was something like where the DApp connects directly to Infura, then this interception will not be applicable.