fix(@cockpit/ens) show error message when ENS registration fails
Turns out Cockpit doesn't handle errors not very well. In the ENS utilities
users can register a subdomain with an address. No matter what is entered
and sent to Embark, Cockpit will always show that the registration was successful.

The server sends an error message correctly, it's just not processed in the
front-end properly. This is due to how our reducers are implemented and in fact,
it looks like the `errorEntities()` reducer doesn't work at all. At least
I couldn't see a case where it would work.

The bottomline for this component is: it tries to read from `errorEntities.ensRecords`
but that state will never be set and in fact, it turns out to be hard to set
as well.

What does get set properly however is the `errorMessage` in the root state.
After trying for a long time to get a proper error state into our store I gave
up and decided to just pass down the `errorMessage` to the ENSRegister component
for now.

Hopefully someone who feels like digging down our reducer rabbit whole will
properly fix this and make all `FAILURE` cases in our reducers/store work.