feat(@embark/blockchain): Add support for blockchain plugins
Blockchain clients can be added as plugins now. Existing `geth` and `parity` clients are still supported, however additional blockchains can used in your DApp as a plugin.

The blockchain plugin API can be used in the following way:
```
this.embark.registerBlockchain("lightchain", require.resolve("./client"), {options});
```
When Embark attempts to start a blockchain process, we first check to see if we have installed a blockchain client plugin, and also check that it's `shouldInit` function resolves to `true`.

If the blockchain client plugin is configured in the DApp's blockchain config for the current environment, we create an instance of the client and then run through some initialisation. Before this PR, the initialisation would attempt to create and unlock node accounts *before* starting up the blockchain process. With this PR, we assume that the CLI for the blockchain client does not support creation and unlocking of accounts from the command line, and instead allow the client to run an init routine before the main process is started.

In addition, we have added a new event, `blockchain:provider:ready`, that is fired once Embark's web3 provider has been configured and set up. This event will run actions configured with `registerActionsForEvent("blockchain:provider:ready")` and wait for those actions to complete before continuing. This allows an opportunity for custom blockchain plugins to perform account management before kicking off the blockchain client process. After the actions are finished, the `blockchain:ready` event will be fired. Because of this, it is important that any blockchain plugins using this event (by default, any that impelement `BaseBlockchainPlugin`) have a valid `shouldInit` method that prevents registration of these actions when the blockchain client is not being used by Embark (ie the blockchain config `client` is not set to the blockchain plugin for the current environment), otherwise the blockchain process may hang.

This PR introduces two new base classes to assist developers in creating blockchain plugins.

`packages/embark-blockchain-process/baseBlockchainPlugin.ts` was implemented as a base class for the blockchain plugin itself. Developers should extend this class when creating blockchain plugins. For example,
```javascript=
import {BaseBlockchainPlugin} from 'embark-blockchain-process';

const BLOCKCHAIN_CLIENT_OPTIONS = {
  defaults: {
    bin: "lightchain",
    networkType: "standalone",
    rpcApi: ['eth', 'web3', 'net', 'debug', 'personal'],
    wsApi: ['eth', 'web3', 'net', 'debug', 'pubsub', 'personal'],
    devWsApi: ['eth', 'web3', 'net', 'debug', 'pubsub', 'personal']
  },
  versSupported: ">=1.3.0",
  versionRegex: "Version: ([0-9]\.[0-9]\.[0-9]).*?",
  name: "lightchain",
  prettyName: "Lightchain (https://github.com/lightstreams-network/lightchain)",
  clientPath: require.resolve("./client")
};

export default class EmbarkLightchain extends BaseBlockchainPlugin {
  constructor(embark) {
    super(embark, BLOCKCHAIN_CLIENT_OPTIONS);
    this.logger = embark.logger;
    this.isDev = embark.config.env === "development";
  }

  async createAndUnlockAccounts() {
    // do node account mgmt here
  }
}
```
Once extended, developers should override the following methods:
1. `protected async createAndUnlockAccounts()`: Creates and unlocks accounts according to the blockchain configuration. Override this method to create a custom account management implementation. If you do not need to create and unlock accounts, override this method in your plugin class, and return immediately. Otherwise, an error will be thrown. This method is fired automatically by the base class in between the `blockchain:provider:ready` and `blockchain:ready` events.
2. `protected shouldInit(blockchainConfig: any)`: Prevents or enables Embark to consider this plugin as a valid plugin at run time. Overridable if needed, the base class implements this method by returning true if the current run time environment's config is using this blockchain plugin's client.

`packages/embark-blockchain-process/baseBlockchainClient.ts` was implemented as a base class for the blockchain client. This class introduces several methods that need to be overridden by the extending class, as can be seen in the [`embark-lightchain` implementation](https://github.com/emizzle/embark-lightchain/blob/master/src/client.js):
1. `public isReady(data: string)`: Determines whether the blockchain client is ready by examining the stdout output of the main command (determined by {@link mainCommand})
2.  `public needKeepAlive()`: check if the client needs some sort of 'keep alive' transactions to avoid freezing by inactivity. If true, allows for txs to be sent at regular intervals via the `devtxs on` console command.
3.  `public getMiner()`: Gets path to a miner class. If a miner is not needed, override this method and log a warning.
4.  `public determineVersionCommand()`: Command to execute in the CLI to determine the blockchain client version.
5.  `public initChain(callback: Callback<null>)`: Fired before the main blockchain command is run, the client should use this method to perform any initialization routine that it needs.
6.  `public mainCommand(address: string, done: (bin: string, args: string[]) => void)`: Main command to run in the CLI that starts up the blockchain client process

`blockchain:provider:ready`: Fired once Embark's web3 provider has been configured and setup. This is fired **before** `blockchain:ready`.