feat(@embark/embark-deployment): Don’t hang on contract deploy fail
Prior to this PR (including with v4), Embark would hang on a deployment error.

After this PR, Embark continues it’s run routine despite a contract deployment error, and reports the error to the user in the console appropriately.

NOTE: this branch is based on `refactor/add-contract-tracking` (which would land in https://github.com/embark-framework/embark/pull/1743). If #1743 lands (will be merged in to `refactor_5_0_0`), I will rebase this branch on top of `refactor_5_0_0`.