fix(@embark/cmd_controller): ensure blockchain module is started before storage
This is crucial as the storage module tries to generate/eval artifacts that
depend on a `web3` object inside the VM scope. The registration of that object
can only be garuanteed when the `blockchain:started` event has been fired.

This commit ensures these modules are not spinned up in parallel as we're running
into race conditions.