feat(@embark/embarkjs): introduce Blockchain.isAvailable() API
This is in order to be consistent with existing `isAvailable()` APIs on
the `EmbarkJS` module.

Another reason this API is needed, is because we need a way to fire
`EmbarkJS.onReady()` even if no blockchain is available. Prior to this
commit, `EmbarkJS.onReady()` was purely relying on an API on `EmbarkJS.Blockchain`
that would break if the blockchain feature was disabled, causing
`EmbarkJS.onReady()` to never fire.