fix(@embark/ens): fix broken test due to async API
As part of https://github.com/embark-framework/embark/pull/2110/commits/8ed5376533b5a165e478f5f12ca74230c33beb7a we made `ENS.getEnsConfig()`
asynchronous, hence it needs a callback. This change was not done in one of the
ENS tests, resulting in the test to fail. The test mocked that particular API,
turning it into noop, which is why the test kept timing out.

This commit fixes the test by not mocking the API and adding the needed configuration
to the environment.