ci: revise "ci:full" script to run some steps collectively
This results in faster completions on Azure Pipelines. For an individual
build (OS/Node.js combination) it shaves off ~10 or more minutes, depending on
the OS.

Once the `lint` and `test` steps have collective implementations we should be
able to shave another couple of minutes off of the completion times.

The steps that take the longest are `yarn install` (~4 minutes), collective
`typecheck` (~4 minutes), and `ci:dapps` (~3 minutes).