build: implement a script to fix intra-monorepo dep range specifiers in rebased branches
Range specifiers for intra-monorepo dependencies can, because of nightly or
stable releases, get out of sync between PR branches and `master`. In many
cases, what then happens is that yarn can't satisfy the dep range with the
version in the monorepo, so it installs from the registry and updates
`yarn.lock` accordingly. We have a `check-yarn-lock.js` script (`yarn cylock`)
that detects the situation and prevents CI from passing. However, to date the
problem has to be fixed manually.

Implement a script that fixes the problem by scanning all the `package.json`
files and updating mismatched ranges.

It's best to run the script *after* rebasing against `master`.